feat: [PQ-223] moved files into new directories.
