+++ behave.ini
@@ -1,5 +1,5 @@
 [behave]
-paths = src/bdd
+paths = src/bdd/wisp
 format=json
 logging_clear_handlers=yes
 logging_filter=-suds
